Move declarations around
Close #24
